Global disable scheduler _sleep instead of mocking.
It is becoming increasingly difficult to predict when
a call to TaskRunner._sleep needs to be scripted, and is probably an
implementation detail which shouldn't be exposed to unit tests anyway.

This change defines a global which prevents sleeping in calls to _sleep
and toggles that global in HeatTestCase setUp.

Change-Id: I98dc88c9d120c409d5720895bb4fb625a1f12991