Small tweaks to recreation of remote errors
The patch adds heat common exceptions to the list of trusted exceptions
that can be re-spawned across RPC calls (this is the default behavior
for all OS components). The patch adds the corresponding package to the
declaration of the allowed_rpc_exception_modules as well as to the conf
sample.

Change-Id: Ib89d0bd4dd8f5240a9f3e973b8a04d0a5bfe2193