Replace rackspace PublicKey property with key_name.
This fetches a keypair that is registered with nova and
uses the public key, instead of having to specify the entire
public key as a launch parameter.

Tested to work on instances launched in DFW.

This shares the keypair lookup method with AWS::EC2::Instance but
adopts the property name of the underlying nova API. This property name
will align with the yet-to-be-written OS::Nova::Server.

Change-Id: I2dceb93a4544e5f99d246233f240fe99667efadf