Assign rackspace server name from physical_resource_name.
The default implementation of physical_resource_name() assigns a name
which includes some context and randomness.

physical_resource_name() is overridden to optionally allow the 'name'
property to specify the physical name. The same technique is used in
OS::Swift::Container, and 'name' matches the property in the underlying
Nova API.

Change-Id: I5769513a4461d5067dca516d0f03e02bb444f379