Make Volumes work with the default AvailabilityZone
The most common way to specify an AvailabiltyZone for a volume is to grab
it from the Instance it is going to be attached to:

    "AvailabilityZone" : {"Fn::GetAtt": ["WikiDatabase",
                                         "AvailabilityZone"]},

But the first unit tests to check the value passed to Cinder for the AZ,
introduced in 134b11ccd145f982b7173bc9a5d0f1bc3e8eef8c, just forced the
value instead of obtaining it from the Instance.

Add a unit test to check the common case, and fix the resulting error (an
empty string is returned instead of None).

Fixes bug #1207416

Change-Id: I83ea2f3f0c417eb9e97abdf27a9eaffd441cb070