Tests for s3 and swift resources were extended
According to the description of the bug, the problem
is in the mechanism of dependencies. That is an internal
function Ref is not working properly. To repeat it, check
was added to the test templates (s3_test, swift_test). Thus,
the bug has not been reproduced.

Also, the method FnGetRefId in s3 and swift resources,
despite its name, returns the resource name in the format
'stack_name-resource_name-resource_id'. This is achieved
by the method physical_resource_name, that implemented in
superclass resource.Resource.

Bug 1202987

Change-Id: I0f85c39e10e8b455f93fdd768cd5e6ee14497277