Allow Parameters to set defaults for TemplateResource
Although it is theoretically possible to infer default values for the
Properties of a TemplateResource from the Parameters of the provider
template, this presents a number of conversion issues that are presently
unhandled.

Instead, pass parameter values only for properties that are supplied.  For
the others, allow the parameter defaults in the template to be used.

Change-Id: I2ba49ffeee06c3a272694c18b3875f843ade1e6c