Remove unnecessary aws_* from user_creds
We currently store aws_creds, but it's never used in any of the clients
which consume the stored context.  It makes sense to remove it from the
DB, but leave it in heat/common/context.py; then the credentials used for
real requests *to* heat (authenticated via the ec2token middleware) are
obtainable in the context, but we don't populate them in context to be
used in client connections for requests going *from* heat.

Fixes bug #1215850

Change-Id: I7a75c70beb0318a1e83ac0ab45d13802999d1282