Don't delete failed instances in InstanceGroup
This gets rid of the code to delete failed instances in
InstanceGroup so that administrators can debug what happened
if any instances fail to start. It also allows the resize operation
to raise an error, so the InstanceGroup should be in a FAILED state
after a failed resize operation.

fixes bug 1214007
fixes bug 1215140

Change-Id: Id508195c847f09dd986e450f35876ff574ff7673