Provide user control for maximum nesting depth
In the case of an infinitely recursing stack, Heat will continue to
keep nesting stacks until python's maximum stack recursion depth
is reached. By this point a lot of memory and time may have been
spent parsing/loading/etc.

The default of 3 is enough to deploy a stack of stacks of stacks. It
can be raised by deployers if there is a desire to do larger and more
complicated stacks, at the cost of more resource utilization.

Fixes bug #1214239

Change-Id: Ic492ef84b94b1f715c49eef7e1794a486fb8182f