Don't use a query for stack_get()
When getting a database record by ID, it is much more efficient to get() it
directly than to perform a query or the database to obtain the ID (again)
and then get it.

Since stack_get() returns only a single record that can be trivially
checked for deletion in code, do this rather than performing a query with a
filter.

Change-Id: I4bb3ea1a9c4ae928fe0f5bce5c02b842a3f1ab4f