Implement the SubnetId property in the Instance resource
If the SubnetId parameter is given in the template, the handle_create method
in the Instance resource verifies the subnet in quantum and configures the
NetworkInterfaces for the subnet. If no NetworkInterface is specified, a port
is created for the subnet and the nic is configured to use that port.

Porting this fix into stable/grizzly

(cherry picked from commit 58d88f64d79ecc336a15e5898b1de1efd605f112)
Conflicts:
    heat/engine/resources/autoscaling.py
    heat/tests/test_instance.py
    heat/tests/test_instance_network.py

Change-Id: I92da0800565392004c4f75ad186f67b638df2df1
Fixes: bug #1163952