Conform stack resource error so it gives a 4xx
This exception was added and landed while the change to use
RequestLimitExceeded was still in progress. Giving a 5xx error when the
user has caused the problem directly is not what we want.

We leave the exception in place to allow the shared error message. I
also have hopes that we will be o-k with extra exceptions and just
extend RequestLimitExceeded when we have a similar situation.

Change-Id: Ic9b176625e6ad2b8c42c420dab79984d134304e5
Partial-Bug: #1230191