Begin 3 chapters for Orch API Ref Manual
- The original heat wadl (heat-api-1.0.wadl) and the samples have now
been deleted from this repo. The wadl has been renamed and migrated
to the following location along with the samples:
https://github.com/openstack/api-site/blob/master/api-ref/src/wadls/
orchestration-api/src/v1/orchestration-api.wadl

- Inserted api-ref/src/wadls/heat-api/src/README.rst file in repo
that specifies the new wadl location.

- Updated pom file to latest version of clouddocs-maven-plugin.

- Currently clouddocs-maven-plugin 1.11.0 has a bug which creates
new sections for the request and response examples for each API
call. This will be fixed in an upcoming release of the plugin.

- Several remarks are inserted in the manual text and marked with
security="writeronly" so they do not appear in the output. You
are welcome to provide any answers and feedback you may have for
these remarks.

- Fix title for commit to resolve H803 build exception.

- Fix extra spaces in file /wadls/heat-api/src/README.rst, and
wrap lines in commit message at less than 70 characters.

Change-Id: I478b394268f64b41317ab77fff6d6366ddd7efa1