Handle NetworkNotFoundClient and PortNotFoundClient
neutronclient has changed is behavior (recently?) in that it now
raises PortNotFoundClient for missing ports and NetworkNotFoundClient
for missing nets. These exceptions extend NeutronClientException but
do not set the status_code to 404 so the existing resource 404
checking has stopped working.

This change handles both 404 exceptions and the new exceptions
so that it continues to work with different neutron releases.

Change-Id: Id06e04d4b98e62b8062b119e959634fd61c8beb5
Closes-Bug: #1234924