Rolling update support for Instance/AutoScalingGroup
This is the last part of the series to implement support
for AutoScaling UpdatePolicy.

Modified handle_update of InstanceGroup and AutoScalingGroup to
support rolling update if there is an update policy and when
there is a property difference in LaunchConfigurationName. A new
replace method is created to handle replacement of instances in
batches. Replacement of instances occurs prior to resize. The
_lb_reload is called appropriately for each batch. The resize
method is refactored to support underlying changes in
_create_template.

Change-Id: I8113a964fb6eb0ec16c2e3766d2771e64e1cf248
Implements: blueprint as-update-policy