Fix possible race issue with test_engine_service
There's the potential for racy behavior in test_show_watch, because
the result of show_watch(self.ctx, watch_name=None) is all watches
visible to the user, not only those for a specific stack.  So it's
possible that the result length will be more than two, and since the
watch name is also not unique, we could see spurious test failures.

So make the test assertions more specific, and avoid reusing the same
watch name in test_show_watch_metric.

Change-Id: Ibb1cbae04dee4b6c4ada44336143fd122a3cb53f
Closes-Bug: #1237324