Raise error if instances are created with names > 63 chars
The maximum hostname on Linux is 64 characters.  DHCP is limited to
this host maximum.  Unfortunately Nova allows the creation of servers
with names > 63 characters in length.  This patch is a short-term
solution to improve the failure condition, so we draw attention to
the problem, enabling the user to work around with shorter names.
A subsequent patch will be required which mangles names such that we
truncate or compact names automatically.

Closes-Bug: 1238272
Change-Id: I2114135b7f4b649a906c3adaea1869d1a1290a3b