Remove a buggy method in sqlalchemy.api
watch_data_delete tries to retrieve session object from a result set instead of
managed objects. Further, from the name of the method, the intention seems to
remove watch data associated with a given watch rule, but it would actually
remove all watch_rules having the given watch_name. Though it would not succeed,
if there is any associated watch data.

Since this method is not referenced anywhere in the entire project, it might be
better to just remove it for now.

Change-Id: I7a66f336a2a33b4a941df85910f29c3cfb2c816f