RS LoadBalancer should return its ID as Ref
Previously the Rackspace::Cloud::LoadBalancer resource returned the name as
its reference ID. This patch reverts this to the default resource behavior
of returning the resource ID if available, and then falling back to the
name. Otherwise there is no way to access the ID of the load balancer,
which is necessary for other Rackspace APIs, such as Rackspace Auto Scale.

Closes-Bug: #1241194
Change-Id: Ib5ca296b338e912daa8c3dc1c71fc3e994a0156e