Allow RS LoadBalancer resource to have no nodes
It should be possible to create Rackspace load balancer with no nodes
now. There was an incidental bug in the way that the resource constructed
the node list that would prevent this.

Note that it won't be possible to actually use this to create a nodeless LB
until a new release of pyrax is made, but this change should be harmless enough
to be merged before that.

Change-Id: Icb9b50899cc13094e295280f89090def30cb4155
Closes-Bug: #1241221