Allow plugins to be specified as a cloud_backend
Currently, the cloud_backend configuration option can only use modules
imported in-tree before clients.py is imported. This does not work for
plugins because they are dynamically imported in
heat/engine/resources/__init__.py after clients.py is imported. Change
the cloud_backend module to be specified when the client class needs
to be initialized rather than when clients.py is imported. This change
requires a class to be specified for the cloud_backend rather than a
module name.

Change-Id: I8f2b93bfb6a9585d1e7dc52d70092105b6e1bfaa
Closes-Bug: 1242870