Add rebuild to OS::Nova::Server
Nova has support to rebuild an instance so that it keeps the same
instance id and address. This is particularly useful for doing image
based updates to servers without incurring additional cost to create new
servers and then delete the old ones.

We still default to replacing, and add a new policy called 'REBUILD' for
the times where we want a resource to be rebuilt on image updates rather
than replaced.

blueprint nova-server-rebuild

Change-Id: If0033f52912e939048b5ebc1ee2e30f67d225f34