Repeat Ceilometer alarm actions by default
The policy actions assume repeat_actions=true.

For the situation when cpu is high even when a new server is added.
Here we would expect yet-another server to be added but what currently
happens is the alarm will be stuck in the alarm state, but with no
further actions.

We need a default that it sane.

Closes-bug: #1242168
Related-bug: #1223347
Change-Id: I4c6a084076a0cc01ebd0d7431d3fac559161e947