Ensure autoscaling actions occur for percentage adjustment
Fixes bug 1251007

Previously if the AdjustmentType was set to PercentChangeInCapacity,
then depending on the choice of the instance group MinSize and the
scale-up policy ScalingAdjustment, no autoscaling actions may occur
even when the under-scaled alarm fires.

This situation occurred if:

   (MinSize * ScalingAdjustment / 100.0) < 1.0

in which case the group never get gets scaled up even if the under-
scaled alarm stays in the alarm state forever.

Now we follow the rounding rules used by AWS Autoscaling:

  http://docs.aws.amazon.com/AutoScaling/latest/DeveloperGuide/as-scale-based-on-demand.html

Change-Id: I79e5e743de1944ace5ef2c4dbcb76c9cab7262bb