Add heatclient to available clients
The first resource to use this will need to ensure
that this method has test coverage. There appears to be
no convention for writing separate unit tests for these
client methods.

This will currently not work with standalone heat as
the heat endpoint will not be in the keystone catalog, see
related-bug: #1255713

partial blueprint: hot-software-config
Change-Id: I325664dcb0ca4d01be41614496c92e65bea52310