Avoid eventlet-unsafe exception handling
Eventlet can context-switch during a log call (at least in principle).
Therefore it is unsafe to use "raise" after this call, because the exception
raised may not be the same one caught by the "except" block, but rather
could be any random exception from some other greenlet.

This is a canonical example of why eventlet is pure evil.

Fixes bug: #1255506

Change-Id: Ieb0eb763c0fd696b82212653ac4909d2bddf6b88