Catch error deleting trust on stack delete
When deleting a stack, it's possible for deleting the trust to fail,
for example if the user deleting the stack is not the user who created
it, or an admin (which raises a Forbidden error), or due to some
other transient error e.g connection to keystone interrupted.

Currently in this case, we fail to mark the stack deleted in the DB
and leave the status "DELETE, COMPLETE", which is misleading.

Change-Id: Ie8e9ea48bc4f44e56ff4764123fcca733f5bd458
Partial-Bug: #1247200