Change ID column of Event table to UUID
If you are using a public Heat you can use the event ID in
'heat event-list' to measure how much activity the service provider
has in their heat installation.

To avoid this, the Id column of the Event table was changed from
autoincremental integer to a uuid.

Added some unit tests in order to ensure that events with integer or
uuid id pass through the API, since we have a migration which will
leave behind integer ids.

Change-Id: Ib3eddb700738cb3667c72977a9a5011209acd6d7
Closes-Bug: #1249159