Allow docker plugin to run w/o deps
The docker plugin tests would not run without the docker lib being
installed even though the client was mocked in the tests. This
patch adds conditional import of the client and registration of
the plugin so that the contrib tests can be run in the gates
without having to hack requirements or add more third party
libs.

Change-Id: Ia114f8ebc0cf0b72af8e6e5d2c3264fb7952d7f9
Partial-Bug: #1251380