wsgi.Resource exception handling to not log errors
In order to not log some exceptions as errors, the
wsgi.Resource.__call__() was modified.
Instead of check if an exception is a HTTPBadRequest or a HTTPNotFound,
now all the exceptions that are not childs of HTTPError are raised
directly, without logging, translation or disguise.
If the exception is a child of HeatException, then the exception is
tranlated but not logged as an error. When this exception reaches the
Fault Wrapper Middleware it is logged as debug.
Some unused, or only use in test, exceptions were removed.

Change-Id: I105eeb74a55dcabc79cd3926dd88da0e19d1e1f1
Closes-Bug: #1231133