SignalResponder store access/secret in resource data
We need to stop relying on the logic in heat_keystoneclient which
assumes there is only one keypair per user, as this assumption
may not remain true in future.

So instead, store the access/secret key in the DB as resource data,
which is probably better from a performance perspective anyway, and
will also allow SignalResponder subclasses to potentially override
the default credentials implementation more easily.

Change-Id: I7952181f30c94848acf358ab34cf17b6281d43ac
Partial-Bug: #1089261
blueprint: instance-users