Prevent tempest from failing upon ActionInProgress
Jenkins occasionally fails in the 'check-tempest-dsvm-postgres-full'
test.  The tempest tests sometimes try to delete the same stack twice
and an ActionInProgress exception is raised.  This results in ERROR
messages in the log file, which causes the tempest test to fail.

The fix is to decorate the StackLock.acquire method with
client_exceptions, which "allows the declaration of expected exceptions
that the RPC layer should not consider fatal, and not log as if they
were generated in a real error scenario."

Closes-Bug: #1261433
Change-Id: I0692b9d5cf03501ba532b40abf4e567134cda057